[LSO-12300] Fix to-api filter in case of non-api request
In case when we had to-api filter in place and we
are handling sqs request we would still be processing to-api filter.

We had additiona check for deletion of partially processed by counting
batch size.

IN case that batch size was != 4 we would not delete messages processes
and they would be retried and de-duplicated.

In case batch-size was exactly 4 we would count 4 fields in map and
delete messages in batch.

Out system montiring will detect such cases and usually in
prod environment batch size i 1.

Change-Id: Ib053a5b023d075a247ea2e29dba4e73d82727642